requirements.txt
requests
python-dateutil
psycopg2-binary
python-dotenv
